Smarter material editing
